r_main.c: remove references to unreleased "passages" feature
this is discussed briefly in Michael Abrash's Black Book, in case
anyone else reads this commit and is curious. it was an alternate
world rendering technique explored during the last month before
Quake shipped, using an optimized (precalculated) version of the
"portals" technique used in other engines of the era.